+++ src/kafka-file/languageservice/kafkaFileLanguageService.ts
@@ -1,8 +1,9 @@
-import { CodeLens, TextDocument, Uri } from "vscode";
+import { CodeLens, CompletionList, Position, TextDocument, Uri } from "vscode";
 import { ConsumerLaunchState } from "../../client";
 import { ProducerLaunchState } from "../../client/producer";
 import { KafkaFileDocument, parseKafkaFile } from "./parser/kafkaFileParser";
-import { KafkaFileDocumentCodeLenses } from "./services/codeLensProvider";
+import { KafkaFileCodeLenses } from "./services/codeLensProvider";
+import { KafkaFileCompletion } from "./services/completion";
 
 /**
  * Provider API which gets the state for a given producer.
@@ -27,41 +28,52 @@ export interface SelectedClusterProvider {
 
 /**
  * Kafka language service API.
- * 
+ *
  */
 export interface LanguageService {
     /**
      * Parse the given text document and returns an AST.
-     * 
+     *
      * @param document the text document of a kafka file.
-     * 
+     *
      * @returns the parsed AST.
      */
     parseKafkaFileDocument(document: TextDocument): KafkaFileDocument;
 
     /**
      * Returns the code lenses for the given text document and parsed AST.
-     * 
+     *
      * @param document the text document.
      * @param kafkaFileDocument the parsed AST.
-     * 
+     *
      * @returns the code lenses.
      */
     getCodeLenses(document: TextDocument, kafkaFileDocument: KafkaFileDocument): CodeLens[];
+
+    /**
+     * Returns the completion result for the given text document and parsed AST at given position.
+     * 
+     * @param document the text document.
+     * @param kafkaFileDocument the parsed AST.
+     * @param position the position where the completion was triggered.
+     */
+    doComplete(document: TextDocument, kafkaFileDocument: KafkaFileDocument, position: Position): CompletionList | undefined
 }
 
 /**
  * Returns the Kafka file language service which manages codelens, completion, validation features for kafka file.
- * 
+ *
  * @param producerLaunchStateProvider the provider which gets the state for a given producer.
  * @param consumerLaunchStateProvider the provider which gets the state for a given consumer.
- * @param selectedClusterProvider the provider which gets the selected cluster id and name. 
+ * @param selectedClusterProvider the provider which gets the selected cluster id and name.
  */
 export function getLanguageService(producerLaunchStateProvider: ProducerLaunchStateProvider, consumerLaunchStateProvider: ConsumerLaunchStateProvider, selectedClusterProvider: SelectedClusterProvider): LanguageService {
 
-    const kafkaFileDocumentCodeLenses = new KafkaFileDocumentCodeLenses(producerLaunchStateProvider, consumerLaunchStateProvider, selectedClusterProvider);
+    const kafkaFileCodeLenses = new KafkaFileCodeLenses(producerLaunchStateProvider, consumerLaunchStateProvider, selectedClusterProvider);
+    const kafkaFileCompletion = new KafkaFileCompletion();
     return {
         parseKafkaFileDocument: (document: TextDocument) => parseKafkaFile(document),
-        getCodeLenses: kafkaFileDocumentCodeLenses.getCodeLenses.bind(kafkaFileDocumentCodeLenses)
+        getCodeLenses: kafkaFileCodeLenses.getCodeLenses.bind(kafkaFileCodeLenses),
+        doComplete: kafkaFileCompletion.doComplete.bind(kafkaFileCompletion)
     };
 }